requirements.txt
Flask==3.1.2
flask-cors==6.0.1
onnxruntime==1.23.2
rembg==2.0.67
pillow==12.0.0
gunicorn==23.0.0
python-dotenv==1.2.1
supabase==2.27.0